UI Ready for Basic Chat Application Needed in it

+++ src/components/Logout.jsx
@@ -1,21 +1,27 @@
 import React from "react";
-import styled from "styled-components";
 import { useNavigate } from "react-router-dom";
 import { BiPowerOff } from "react-icons/bi";
-
-const Logout = () => {
+import styled from "styled-components";
+import axios from "axios";
+import { logoutRoute } from "../utils/APIRoutes";
+export default function Logout() {
   const navigate = useNavigate();
-
-  const handleClick = () => {
-    localStorage.clear();
-    navigate("/login");
+  const handleClick = async () => {
+    const id = await JSON.parse(
+      localStorage.getItem(process.env.REACT_APP_LOCALHOST_KEY)
+    )._id;
+    const data = await axios.get(`${logoutRoute}/${id}`);
+    if (data.status === 200) {
+      localStorage.clear();
+      navigate("/login");
+    }
   };
   return (
     <Button onClick={handleClick}>
       <BiPowerOff />
     </Button>
   );
-};
+}
 
 const Button = styled.button`
   display: flex;
@@ -23,7 +29,7 @@ const Button = styled.button`
   align-items: center;
   padding: 0.5rem;
   border-radius: 0.5rem;
-  background-color: rgb(255, 82, 161);
+  background-color: #9a86f3;
   border: none;
   cursor: pointer;
   svg {
@@ -31,5 +37,3 @@ const Button = styled.button`
     color: #ebe7ff;
   }
 `;
-
-export default Logout;
